# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateLabel": {
        "enum": [
          "A",
          "B"
        ],
        "title": "CandidateLabel",
        "type": "string"
      },
...
      "CandidateStatus": {
        "enum": [
          "PENDING",
          "RUNNING",
          "SUCCEEDED",
          "FAILED"
        ],
        "title": "CandidateStatus",
        "type": "string"
      },
...
      "CreateRepairJobRequest": {
        "additionalProperties": false,
        "description": "The body of ``POST /v1/jobs``. Clients must also send an ``Idempotency-Key`` header.",
        "properties": {
          "plan_mode": {
            "$ref": "#/components/schemas/PlanMode",
            "default": "AUTO"
          },
          "repair_type": {
            "$ref": "#/components/schemas/RepairType",
            "default": "FOOT_CONTACT"
          },
          "selection": {
            "$ref": "#/components/schemas/AnimationSelection"
          }
        },
        "required": [
          "selection"
        ],
        "title": "CreateRepairJobRequest",
        "type": "object"
      },
...
      "DecisionRequest": {
        "additionalProperties": false,
        "description": "The body of ``POST /v1/jobs/{job_id}/decision``.",
        "properties": {
          "choice": {
            "$ref": "#/components/schemas/DecisionChoice"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "time_to_decision_s": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time To Decision S"
          }
        },
        "required": [
          "choice"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "DefectReport": {
        "additionalProperties": false,
        "properties": {
          "algorithm_version": {
            "title": "Algorithm Version",
            "type": "string"
          },
          "detection_config": {
            "$ref": "#/components/schemas/DetectionConfig",
            "default": {
              "consistency_band_m": 0.01,
              "floor_percentile": 5.0,
              "gap_close_frames": 2,
              "height_threshold_m": 0.025,
              "min_run_frames": 6,
              "severity_major_cm": 2.0,
              "severity_minor_cm": 0.5,
              "speed_threshold_m_s": 0.6
            }
          },
          "intervals": {
            "items": {
              "$ref": "#/components/schemas/PlantedInterval"
            },
            "title": "Intervals",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "summary": {
            "maxLength": 1000,
            "title": "Summary",
            "type": "string"
          },
          "worst_interval_index": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worst Interval Index"
          }
        },
        "required": [
          "intervals",
          "severity",
          "summary",
          "algorithm_version"
        ],
        "title": "DefectReport",
        "type": "object"
      },
...
      "ErrorInfo": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "maxLength": 2000,
            "title": "Message",
            "type": "string"
          },
          "node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo",
        "type": "object"
      },
      "EvaluationReport": {
        "additionalProperties": false,
        "properties": {
          "evaluator_status": {
            "$ref": "#/components/schemas/EvaluatorStatus"
          },
          "objective_ranking": {
            "items": {
              "$ref": "#/components/schemas/ObjectiveRank"
            },
            "title": "Objective Ranking",
            "type": "array"
          },
          "recommendation_reason": {
            "default": "",
            "maxLength": 2000,
            "title": "Recommendation Reason",
            "type": "string"
          },
          "recommended": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "scoring_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Scoring Weights",
            "type": "object"
          },
          "visual": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/VisualEvaluation"
            },
            "title": "Visual",
            "type": "array"
          }
        },
        "required": [
          "objective_ranking",
          "evaluator_status"
        ],
        "title": "EvaluationReport",
        "type": "object"
      },
...
      "EventPage": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/JobEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "next_seq": {
            "description": "Pass as after_seq on the next poll",
            "minimum": 0.0,
            "title": "Next Seq",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "next_seq"
        ],
        "title": "EventPage",
        "type": "object"
      },
...
      "JobStatus": {
        "enum": [
          "PENDING",
          "RUNNING",
          "AWAITING_DECISION",
          "APPLYING",
          "COMPLETED",
          "REJECTED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "PlanMode": {
        "enum": [
          "AUTO",
          "DETERMINISTIC"
        ],
        "title": "PlanMode",
        "type": "string"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "An RFC 7807 problem detail, extended with a machine-readable ``code``.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RepairCandidate": {
        "additionalProperties": false,
        "properties": {
          "algorithm_version": {
            "default": "foot_contact/1",
            "title": "Algorithm Version",
            "type": "string"
          },
          "artifacts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ArtifactReference"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "candidate_id": {
            "pattern": "^[0-9a-f]{16}$",
            "title": "Candidate Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "label": {
            "$ref": "#/components/schemas/CandidateLabel"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "parameters": {
            "$ref": "#/components/schemas/CandidateParameters"
          },
          "status": {
            "$ref": "#/components/schemas/CandidateStatus",
            "default": "PENDING"
          }
        },
        "required": [
          "candidate_id",
          "label",
          "parameters"
        ],
        "title": "RepairCandidate",
        "type": "object"
      },
      "RepairJob": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RepairCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "defect": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DefectReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9_.:-]{8,128}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "job_id": {
            "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
            "title": "Job Id",
            "type": "string"
          },
          "last_event_seq": {
            "default": 0,
            "minimum": 0.0,
            "title": "Last Event Seq",
            "type": "integer"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepairPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan_mode": {
            "$ref": "#/components/schemas/PlanMode"
          },
          "repair_type": {
            "$ref": "#/components/schemas/RepairType"
          },
          "selection": {
            "$ref": "#/components/schemas/AnimationSelection"
          },
          "skeletal_scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkeletalScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "idempotency_key",
          "repair_type",
          "plan_mode",
          "selection",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "RepairJob",
        "type": "object"
      },
...
      "RepairType": {
        "enum": [
          "FOOT_CONTACT"
        ],
        "title": "RepairType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "AI plans and evaluates; deterministic code executes; the animator decides. Every error body is an RFC 7807 Problem carrying a machine-readable `code`.",
    "summary": "Localized animation repair: measure, propose, compare, apply.",
    "title": "Kinesis API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "description": "Validates the selection against the uploaded scene's inventory (armature, bones, frame range) and schedules the repair DAG. A replay with the same `Idempotency-Key` and body returns 200 with the existing job. The same key with a different body returns 409 IDEMPOTENCY_CONFLICT.",
        "operationId": "create_job_v1_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_.:-]{8,128}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepairJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairJob"
                }
              }
            },
            "description": "Idempotent replay: the existing job"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Create a repair job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Get a job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Cancel a non-terminal job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/candidates": {
      "get": {
        "operationId": "list_candidates_v1_jobs__job_id__candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RepairCandidate"
                  },
                  "title": "Response List Candidates V1 Jobs  Job Id  Candidates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "List repair candidates",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/jobs/{job_id}/decision": {
      "post": {
        "description": "Allowed only in AWAITING_DECISION (otherwise 409 INVALID_STATE). Choosing A or B schedules the non-destructive apply (status APPLYING). REJECT_ALL moves the job to REJECTED. Choosing a FAILED candidate returns 422.",
        "operationId": "post_decision_v1_jobs__job_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Record the animator's choice",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/defect": {
      "get": {
        "operationId": "get_defect_v1_jobs__job_id__defect_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefectReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Deterministic defect report (409 NOT_READY until analysis completes)",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/evaluation": {
      "get": {
        "operationId": "get_evaluation_v1_jobs__job_id__evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Combined objective + model evaluation",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "list_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Poll job events after a sequence number",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}